+++ src/kernel/coinstats.cpp
@@ -2,7 +2,7 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include <node/coinstats.h>
+#include <kernel/coinstats.h>
 
 #include <coins.h>
 #include <crypto/muhash.h>
